test(l1): fix hive ReOrg tests by requesting headers from newToOld (#4676)
**Motivation**

The following tests in the `engine-withdrawals` hive test suite:

- Withdrawals Fork on Block 1 - 8 Block Re-Org, Sync (Paris) (ethrex)
- Withdrawals Fork on Block 8 - 10 Block Re-Org Sync (Paris) (ethrex)
- Withdrawals Fork on Canonical Block 8 / Side Block 7 - 10 Block Re-Org
Sync (Paris) (ethrex)
- Withdrawals Fork on Canonical Block 8 / Side Block 9 - 10 Block Re-Org
Sync (Paris) (ethrex)

failed because when syncing we always request the block headers from our
`current_head` on (OldToNew), but in these tests the request goes to a
node that's on a sidechain that didn't have our `current_head` so the
request returned no headers and syncing never finished.

**Description**

In this pr I tackle the case described by having a first request of
headers from NewToOld from the `sync_head`, and then trying to sync,
this is all covered in a new function `synced_new_to_old`. If syncing
fails because we have no common ancestor within the obtained headers we
go back to request headers as we did before, from OldToNew from our
`current_head` and then try to sync as usual.

Closes #4577 and partially solves #2126 as a provisional solution until
#4717 is solved.

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>