refactor(l1): throw error on wrong version of eth capability (#4788)
**Motivation**

Stop panicking when decoding a message with a wrong version of
capabilities `eth68` and `eth69`.

**Description**

This pr has the `decode` method for both capabilities throw error in
case the eth version of the message is wrong.
The only instance where this can happen is if the other peer isn't
respecting the negotiated capabilities. This happens as seen in issue
#4760.
[Here](https://github.com/lambdaclass/ethrex/blob/05e0de5aaa40df9ec7088f7a20a47720c56cd350/crates/networking/p2p/rlpx/connection/server.rs#L676-L677)
we can see the eth version negotiated can only be either 68 or 69.

Closes #4760