feat(l2): custom native token  (#4622)
**Motivation**

We want to be able to start the L2 with the native token to be one
selected by the user. This means the currency on which L2 balances and
fees are paid are actually ETH or an ERC20 selected.

**Description**

* Added a new `native_token_l1_address` option to `DeployerOptions` in
the deployer, allowing configuration of the L1 address of the L2 native
token (ETH or ERC20). This value is now written to the `.env` file after
deployment.
* Updated contract deployment and initialization to pass the
`native_token_l1_address` to the bridge contract, and updated the bridge
initializer signature to accept this new argument.
* Refactored the Solidity `CommonBridge` contract to support deposits
and withdrawals of either ETH or a configurable ERC20 as the native
token. And updates the `deposit` and `claimWithdrawal` logic
accordingly.
* Updated integration tests to support the new native token
configuration, passing the `native_token_l1_address` through test
helpers.
* Added a new contract binary in the fixtures `NativeTokenTest.bin` is
the same as the `TestToken.bin` but with more minting for enabling the
deposits as they use much more. And also to not collide with the
contract that is deployed for the integration tests.

**How to test**

We need to deploy the L1 token contract before starting the L2. We need
two terminals

Start L1
```
cd crates/l2
make rm-db-l1 init-l1
```

Deploy the L1 token
```
rex deploy $(cat fixtures/contracts/ERC20/NativeTokenTest.bin) 0 0xbcdf20249abf0ed6d944c0288fad489e33f66b3960d9e6229c1cd214ed3bbe31
```

Add in the `deploy-l1` Makefile rule under `crates/l2` the line
`--native-token-l1-address 0xb4B46bdAA835F8E4b4d8e208B6559cD267851051`

Then
```
cd crates/l2
make rm-db-l2 deploy-l1 init-l2
```

---------

Co-authored-by: Tomás Paradelo <112426153+tomip01@users.noreply.github.com>
Co-authored-by: Tomás Paradelo <tomas.paradelo@lambdaclass.com>