refactor(levm): use `AccountStatus` for managing status of accounts (#4833)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Start using the `AccountStatus` enum that we created some time ago in
order to know the state of an account at a given moment.

**Description**

- We now don't need the hashmap of `destroyed_accounts` in the
GeneralizedDatabase because each account will keep track of their state.
- Improved documentation for cases in which an account storage has to be
removed but not the account. We had in our mind just one case but we
weren't contemplating the case in which an account is destroyed but then
sent value (without being created again). In this scenario the account
won't be removed because it's not empty, that's why it falls into the
`DestroyedModified` category.
- Keep track of accounts that have been modified so that we can skip
processing unmodified accounts when getting state transitions.

Closes #3948 (issue related to the `destroyed_accounts` part)