feat(l2): enable command flags for `l2 --dev` (#4880)
**Motivation**

When running `ethrex l2 --dev` the sequencer flags are being ignored. We
want to be able to configure the L2

**Description**

- Add a new `populate_with_defaults` function for L2 options. This
function assigns default values to any `Option<T>` flags that are not
provided when running `l2 --dev`, instead of leaving them as None.
- Update `default` implementation for `ProofCoordinatorOptions` to set
`proof_coordinator_qpl_tool_path` as the one in the cli