chore(l1): remove snapshot dirs when they are no longer used (#4815)
**Motivation**

Snap sync generates some intermediate files to limit the amount of
memory used during the process. These are filled with the leaves
downloaded during the first step of snap sync, before healing. These are
large files that remain after the node is done using them. This pr aims
to delete them, and to not launch a run if the previous files are in
disk.

**Description**

- Deletes snapshot files for account state, storage slots and the
temporary rocksdb db used for faster insert.
- Exits the process if we find the files in disk. Example of the log in
that scenario.\

`ERROR ethrex_p2p::sync: One of the folders used for temporary leaves
during snap is still used. Delete them in
/home/admin/.local/share/ethrex`

---------

Co-authored-by: Mario Rugiero <mrugiero@gmail.com>
Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>