refactor(l1,l2): separate binaries (#4829)
**Motivation**

We aim to reintroduce the L2 feature flag making so that ethrex has 3
available binaries, the L1 execution client, the L2 client and the
L2+GPU client. This is done for the following advantages:
- Less crates needed for the L1 as the UI, L2, zkVMs crates are
excluded.
- L1, L2, y zkVMs are compiled with the optimum crates for performance
and their needs.
- The L1 user only downloads a client without unneeded functionality.

**Description**

- Added feature flag l2 to ethrex to compile l2 dependencies separately
- Added feature flag l2-sql to ethrex to launch l2 with the SQL database
for the rollup
- Modified CI to create 2 tar images of ethrex for testing, ethrex:main
and ethrex:main-l2
- Modified lint commands in makefile to lint l1 and l2 separately. L1
can't handle the workspace flag, due to some l2 crates in the workspace.

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Francisco Xavier Gauna <francisco.gauna@lambdaclass.com>