fix(l1, l2): cargo test compilation at workspace level (#4916)
**Motivation**

`cargo test` at the workspace level or in `cmd/ethrex` is not compiling.
This is because `ethrex-l2-rpc` is a `dev-dependencie`.

This issue uncovered a bigger issue, that is the existence of `benches`
inside `cmd/ethrex`.

**Description**

This PR solves the compilation error by moving benchmarks to a separate
crate, `benches`. It also removes an unsued benchmark.