feat(l1): running some GenServers in dedicated threads (#4925)
Currently all our GenServers run on the same thread. Many of those are a
single process running forever with the application, and may benefit to
be run on their own dedicated thread.

This PR moves to their own threads the `DiscoveryServer`,
`PeerTableServer` and `Initiator`.

It also improves `PeerConnection` logging as it moves many `debug` lines
to `trace`