refactor(l1): remove `Mutex` from profiling metrics (#5031)
**Motivation**

Having a centralized `Mutex` in the profiling metrics could potentially
make code slower, the more functions we instrument. Removing it reduces
the noise we have in our measurements.

**Description**

This PR replaces the singleton `Mutex` by using the internal `RwLock`
that spans already have for layers to store things. Disabling metrics
doesn't seem to increase performance in a noticeable manner, so this
shouldn't increase performance.

---------

Co-authored-by: Edgar <git@edgl.dev>