fix(l1): track requested pooled transactions (#4962)
**Motivation**
Enable hive p2p test BlobViolations
This test behaves by first sending a `NewPooledTransactionHashes`
message, and then responding to our node's `GetPooledTransaction` with a
`NewPooledTransactions` message that contains transactions with sizes &
types that don't match the ones advertised in the first message. As we
were not keeping track of the received `NewPooledTransactionHashes`
messages, we were not able to validate the incoming transactions. This
PR fixes this problem and re-enables the test
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Track incoming `NewPooledTransactionHashes`
* Enable hive p2p `BlobViolations` test
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Part of #4941