refactor(l1): remove unnecessary NodeRecord mutex (#5067)
During `init_l1`, we instanciate a `NodeRecord` named
`local_node_record`
that is used in a few places.
A deep copy is passed to the RPC API via a call to `init_rpc`.
A shallow copy (`Arc::clone`) is passed to the P2P layer via a call to
`init_network`.
Then the original instance is returned to the caller, which passes it on
to `shutdown_process` at the end of the program lifetime.

The following was observed by inspecting the code:
- Only one of `NodeRecord`'s methods mutates it, and it's never called
  (deleted without errors);
- The only other method receiving a mutable reference doesn't mutate it,
  as shown by the fact removing the `mut` caused no errors either;
- `P2PContext` doesn't actually use `local_node_record`, it was deleted
  without issues;
All of this means we can simply share a regular instance and refer to it
by regular reference, avoinding any locking or indirection.

The sequence number is currently only updated on startup, by
initializing with an already incremented seqno.

No perf improvements are expected, but the code is slightly cleaner
without it.

**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number