feat(l1,l2): setting avx2 flags for compilation and adding features asm to crates (#5189)
**Motivation**

We want to take advantage of AVX2 processor features, and we are using
`asm` features of dependencies in case it's possible.

**Description**

This PR sets compilation flag with AVX2 with the configuration:

```
target-cpu=x86-64-v3
```

At the same time, it changes Cargo.toml files in the following way:
* `c-kzg`: no default features, and add `std` and
`ethereum_kzg_settings`.
* `ark-ff`: with `asm` feature.
* `sha3`: unified in Cargo.toml of the workspace and `asm` feature
added.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number